Corrected the dashboard bug
Previously, sometimes, the time spent for each file to be
read was different, and the name vector was defined before
the reading. This way, the labels of the charts sometimes were
on the wrong places. Now, if the time to read varies, the name
axis will vary too and be correct.